+++ cmd/psweb/telegram.go
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	"peerswap-web/cmd/psweb/config"
+	"peerswap-web/cmd/psweb/internet"
 	"peerswap-web/cmd/psweb/ln"
 
 	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
@@ -103,7 +104,7 @@ func telegramStart() {
 				telegramSendMessage(t)
 			case "/version":
 				t := "Current version: " + version + "\n"
-				t += "Latest version: " + getLatestTag()
+				t += "Latest version: " + internet.GetLatestTag()
 				telegramSendMessage(t)
 			default:
 				telegramSendMessage("I don't understand that command")